OOP Physics This is a merge with the Task 3. Branch - the scripted physics are now changed to OOP.
Add SphereBoundingBox collision with SphereBoundingBox
Add Key-Callback for the ESC Key to close the window
Change float to double (we seem to need double precision)

+++ include/cube.hpp
@@ -7,8 +7,8 @@ class Cube : public Drawable {
 public:
 
   Cube();
-  Cube(float origin_x, float origin_y, float origin_z);
-  Cube(float origin_x, float origin_y, float origin_z, unsigned char colors);
+  Cube(double origin_x, double origin_y, double origin_z);
+  Cube(double origin_x, double origin_y, double origin_z, unsigned char colors);
   ~Cube();
 
   virtual void draw();
